Create an app file
